% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{check_grad}
\alias{check_grad}
\title{Sum of the observation-wise gradient contributions to the fitted model.}
\usage{
check_grad(object)
}
\arguments{
\item{object}{an object of class \code{gamtree}.}
}
\value{
Returns a matrix with the sum of the observation-wise gradient
contributions, with a row for each terminal node of the tree and a column
for each coefficient.
}
\description{
Computes the sum of the observation-wise gradient contributions to the
fitted model in each of the terminal nodes. The sums should be reasonably
close to zero.
}
