% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{coef.gamtree}
\alias{coef.gamtree}
\title{Extract coefficients from a GAM tree.}
\usage{
\method{coef}{gamtree}(object, which = "fixed", ...)
}
\arguments{
\item{object}{an object of class \code{"gamtree"}.}

\item{which}{character. Either \code{"fixed"} (default) or \code{"random"}, 
indicating that fixed- or random-effects coefficients should be returned,
respectively.}

\item{...}{further arguments to be passed to \code{\link[lme4]{fixef.merMod}} or 
\code{\link[lme4]{ranef.merMod}}.}
}
\description{
\code{coef.gamtree} extracts fixed- or random-effects coefficients from 
a GAM tree.
}
