% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eco}
\alias{eco}
\title{Example dataset of light-response curves}
\format{
## `eco`
A \code{data.frame} with 628 observations of 5 variables: 
\describe{
  \item{Species}{categorical partitioning variable, indicator for specie (signal).}
  \item{PAR}{continuous predictor variable for the node-specific model (signal).}
  \item{Pn}{continuous response variable.}
  \item{Specimen}{identification number of individual plants.}
  \item{noise}{artifically generated continuous covariate (noise).}
}
}
\source{
{https://stackoverflow.com/questions/37037445/using-mob-trees-partykit-package-with-nls-model}
}
\usage{
data(eco)
}
\description{
The data comprises light-response curves, which describe the relationship between photosynthetically 
active radiation (\code{PAR}) and photosynthetic rate (\code{Pn}). 
Observations are repeated measures on the same plants. Variable `Specimen` provides an identifier 
for individual plants. Dataset was copied from 
https://stackoverflow.com/questions/37037445/using-mob-trees-partykit-package-with-nls-model 
For illustrating the funtionality of package gamtree, two variables 
were added: \code{Specimen}, which identifies individual plants, and \code{noise} 
an artificially generated continuous covariate that is pure (gaussian) noise.
}
\examples{
summary(eco)
}
\keyword{datasets}
