% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{fitted.gamm4}
\alias{fitted.gamm4}
\title{Internal function for extracting fitted values from MOB-based GAM trees.}
\usage{
\method{fitted}{gamm4}(object, ...)
}
\arguments{
\item{object}{an object of class \code{gamm4}.}

\item{...}{currently not used.}
}
\description{
\code{fitted.gamm4} extract fitted values from objects of class \code{gamm4}.
}
