% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{fixef.gamtree}
\alias{fixef.gamtree}
\title{Extract fixed-effects coefficients from a GAM tree.}
\usage{
\method{fixef}{gamtree}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"gamtree"}.}

\item{...}{further arguments to be passed to \code{\link[lme4]{fixef.merMod}}.}
}
\description{
\code{fixef.gamtree} extracts fixed-effects coefficients from a GAM tree.
}
