% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{gamtree}
\alias{gamtree}
\title{Recursively partition a dataset based on a (local) GAM, while accounting for 
globally specified GAM terms.}
\usage{
gamtree(tree_form, gam_form = NULL, data, abstol = 0.001,
  verbose = TRUE, mob_ctrl = NULL, s_ctrl = NULL, ...)
}
\arguments{
\item{tree_form}{specifies the response and predictor variable for the 
node-specific GAM (separated by a tilde), and the partitioning variables, 
separated by a vertical bar. By default, a thin plate regression spline 
smooth (as implemented in \code{mgcv::s()}) will be fitted to the local 
predictor variable. Only one local predictor variable can be specified, 
currently. Arguments to be passed to function \code{s()} should be supplied
to argument \code{s_ctrl} (only for the node-specific model, arguments for
global smooths should be included directly in the \code{gam_form}).}

\item{gam_form}{specifies the global model, that is, the smooth and
parametric terms to be estimated globally. The formula should contain 
the response variable, followed by a tilde \code{~}, followed by the 
global smooth and parametric terms to be fitted. In contrast to the 
\code{tree_form}, the global terms can be specified as is customary for
\code{gam()}. If \code{NULL}, no global (only local) parameters will be 
estimated.}

\item{data}{specifies the dataset (must be a data.frame).}

\item{abstol}{specifies the convergence criterion. If the log-likelihood 
values of the full mixed-effects gam, from two consecutive iterations differ 
less than abstol, estimation is halted.}

\item{verbose}{logical. Should progress be printed to the commande line in 
every iteration? If true, the iteration number, information on the 
splitting procedure, and the log-likelihood (with df) value of the fitted 
full mixed-effects gam model is printed.}

\item{mob_ctrl}{a list with control parameters as returned by \code{mob_control}
to be passed to function \code{mob()}. (Note that argument `xtype` is set to 
`data.frame`, by default, and cannot be adjusted.)}

\item{s_ctrl}{a list of arguments to be passed to the locally
fitted \code{s()} function. If \code{NULL}, the default arguments of 
function \code{s()} will be employed. NOTE: argument \code{by} cannot be
specified, as \code{gamtreee} requires the use of this argument.}

\item{...}{additional arguments to be passed to \code{mob()}'s fitting 
function (currently, the only option is the default \code{gam_fit()}).}
}
\value{
Returns an object of class \code{"gamtree"}. This is a list, containing
(amongst others) the GAM-based recursive partition (in \code{$tree}), and the 
fitted full GAM with both local and/or global fitted effects (in \code{$gamm}).
}
\description{
\code{gamtree} recursively partition a dataset based on a (local) GAM, in
which a smooth function (thin plate regression spline) of a single predictor
variable is estimated, while accounting for globally specified smooth 
functions and random effects.
}
