% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinetree.R
\name{plot.splinetree}
\alias{plot.splinetree}
\title{Plotting function for visualization of spline-based (g)lmertrees.}
\usage{
\method{plot}{splinetree}(x, which = "both", fitted = "marginal", ...)
}
\arguments{
\item{x}{fitted object of class \code{(g)lmertree} containing splines specified
by \code{\link{setup.spline}} in the terminal node model.}

\item{which}{character, "both", "tree" or "ranef". Other options are available, see
\code{\link[glmertree]{plot.glmertree}} and \code{\link[glmertree]{plot.lmertree}}, 
but might be less helpful for spline models.}

\item{fitted}{character, "marginal" (default), "combined" or "none". 
Specifies whether and how fitted values should be computed and visualized. 
See \code{\link[glmertree]{plot.lmertree}} or \code{\link[glmertree]{plot.glmertree}}
for further detail.}

\item{...}{additional arguments to be passed to \code{\link[glmertree]{plot.lmertree}}
or \code{\link[glmertree]{plot.glmertree}}.}
}
\description{
\code{plot.splinetree} takes a fitted \code{(g)lmertree} with splines and plots it. 
It is a wrapper for \code{\link[glmertree]{plot.glmertree}} and  
\code{\link[glmertree]{plot.lmertree}}, with critical adjustments for
better visualization of spline models in the terminal nodes.
}
\examples{
st <- splinetree(Pn ~ ns(PAR, df = 5) | Specimen | Species, data = eco, 
               cluster = Specimen)
plot(st)
}
\seealso{
\code{\link{setup.spline}} \code{\link{predict.splinetree}} 
\code{\link[glmertree]{plot.lmertree}} \code{\link[glmertree]{plot.glmertree}}
}
