% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{predict.gamm4}
\alias{predict.gamm4}
\title{Internal function for extracting predictions from MOB-based GAM trees.}
\usage{
\method{predict}{gamm4}(object, newdata, ...)
}
\arguments{
\item{object}{an object of class \code{gamm4}.}

\item{newdata}{an optional \code{data.frame} in which to look for variables with which to predict.
If omitted, the fitted values are used.}

\item{...}{currently not used.}
}
\description{
\code{predict.gamm4} extract predictions from objects of class \code{gamm4}.
}
