% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_splines.R
\name{predict.splinetree}
\alias{predict.splinetree}
\title{Predict method for spline-based (g)lmertrees.}
\usage{
\method{predict}{splinetree}(object, newdata, ...)
}
\arguments{
\item{object}{fitted object of class \code{(g)lmertree} containing splines specified
by \code{\link{setup.spline}} in the terminal node model.}

\item{newdata}{\code{data.frame} with observations for which predictions should be computed.}

\item{...}{additional arguments to be passed to \code{\link[glmertree]{predict.lmertree}}
or \code{\link[glmertree]{predict.glmertree}}.}
}
\description{
\code{predict.splinetree} computes predictions for a fitted 
\code{(g)lmertree} that is based on splines.
}
\examples{
st <- splinetree(Pn ~  ns(PAR, df = 5) | Specimen | Species, data = eco, 
               cluster = Specimen)
predict(st, newdata = eco[1L:5L, ])
}
\seealso{
\code{\link{setup.spline}} \code{\link{predict.splinetree}} 
\code{\link[glmertree]{lmertree}} \code{\link[glmertree]{glmertree}}
}
