% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{print.gamtree}
\alias{print.gamtree}
\title{Print method for a fitted GAM tree}
\usage{
\method{print}{gamtree}(x, ...)
}
\arguments{
\item{x}{object of class \code{gamtree}.}

\item{...}{further arguments to be passed to \code{\link[partykit]{print.modelparty}}.}
}
\description{
Prints the local and/or global terms in a fitted GAM tree.
}
\examples{
## GAM tree without global terms:
gt <- gamtree(Pn ~ s(PAR, k = 5L) | Species, data = eco, cluster = eco$Specimen)
gt ## or: print(gt)
}
