% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{ranef.gamtree}
\alias{ranef.gamtree}
\title{Extract random-effects coefficients from a GAM tree.}
\usage{
\method{ranef}{gamtree}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"gamtree"}.}

\item{...}{further arguments to be passed to \code{\link[lme4]{ranef.merMod}}.}
}
\description{
\code{ranef.gamtree} extracts random-effects coefficients from a GAM tree.
}
