% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinetree.R
\name{setup.spline}
\alias{setup.spline}
\title{Set up splines bases for use with function (g)lmertree.}
\usage{
setup.spline(spline, data, ...)
}
\arguments{
\item{spline}{character vector of length 1, describing the spline basis to be created.
Currently, functions \code{ns} and \code{bs} are supported. See Examples.}

\item{data}{a \code{data.frame} containing the variable referred to in \code{spline}.}

\item{...}{additional arguments to be passed to the function specified in \code{spline}.}
}
\value{
a \code{data.frame} with a many rows and columns as \code{data}. The spline 
consists of \code{df} basis functions, but is contained in a single column named 
\code{spline}, followed by the name of the predictor variable specified.
}
\description{
\code{setup.spline} takes a dataset and spline specification as input, and returns the 
dataset with spline bases added.
}
\examples{
data <- setup.spline("ns(PAR, df = 3)", data = eco)
head(data)
matplot(x = data$PAR[order(data$PAR)], 
        y = data$spline.PAR[order(data$PAR),], type = "l")
data <- setup.spline("bs(PAR, degree = 2, df = 4)", data = eco)
head(data)
matplot(x = data$PAR[order(data$PAR)], 
        y = data$spline.PAR[order(data$PAR),], type = "l")
}
\seealso{
\code{\link{plot.splinetree}} \code{\link{predict.splinetree}} 
\code{\link[glmertree]{lmertree}} \code{\link[glmertree]{glmertree}}
}
