% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_splines.R
\name{splinetree}
\alias{splinetree}
\title{Fit a (g)lmertree using spline-based partitioning.}
\usage{
splinetree(formula, data, family = "gaussian", ...)
}
\arguments{
\item{formula}{A four-part function See Examples below, and
\code{\link[glmertree]{lmertree}} or \code{\link[glmertree]{glmertree}}.}

\item{data}{a \code{data.frame} containing the variable referred to in \code{spline}.}

\item{family}{family specification for \code{\link[glmertree]{glmertree}}. See 
\code{\link[stats]{glm}} documentation for families.}

\item{...}{additional arguments to be passed to function \code{\link[glmertree]{lmertree}} 
(default, i.e., \code{family = "gaussian"})
or \code{\link[glmertree]{glmertree}} (\code{family} other than \code{gaussian}).}
}
\value{
A object of class\code{"splinetree" }and \code{"lmertree"} or \code{"glmertree"}.
}
\description{
\code{splinetree} is a wrapper for functions \code{(g)lmertree} to simplify
fitting, visualizing and predicting spline-based trees.
}
\examples{
 sp <- splinetree(Pn ~ ns(PAR, df = 5) | Specimen | Species, data = eco)
sp
}
\seealso{
\code{\link{plot.splinetree}} \code{\link{predict.splinetree}} 
\code{\link[glmertree]{lmertree}} \code{\link[glmertree]{glmertree}}
}
