% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{summary.gamtree}
\alias{summary.gamtree}
\title{Summary method for a fitted GAM tree}
\usage{
\method{summary}{gamtree}(object, ...)
}
\arguments{
\item{object}{object of class \code{gamtree}.}

\item{...}{further arguments to be passed to \code{summary.gam}.}
}
\description{
Prints a summary of the local and/or global terms in a fitted GAM tree.
}
\examples{
## GAM tree without global terms:
gt1 <- gamtree(Pn ~ s(PAR, k = 5L) | Species, data = eco, cluster = eco$specimen)
summary(gt1)

## GAM tree with global terms:
gt2 <- gamtree(Pn ~ s(PAR, k = 5L) | s(cluster_id, bs = "re") + noise | Species, 
              data = eco, cluster = eco$specimen)
summary(gt2)
}
