% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{VarCorr.gamtree}
\alias{VarCorr.gamtree}
\title{Extract random-effects covariance matrices from a GAM tree.}
\usage{
\method{VarCorr}{gamtree}(x, sigma = 1, which = "terminal", ...)
}
\arguments{
\item{x}{an object of class \code{"gamtree"}.}

\item{sigma}{an optional numeric value used as a multiplier for the standard deviations.}

\item{which}{character. \code{"terminal"} (default) returns (co)variances for 
all terminal nodes, \code{"inner"} returns the (co)variances for all inner (splitting) 
nodes, \code{"all"} returns covariances for all nodes.}

\item{...}{additional arguments to be passed to \code{\link[lme4]{VarCorr.merMod}}.}
}
\description{
\code{VarCorr.gamtree} extracts fixed-effects random-effects covariance 
matrices from the nodes of a GAM tree.
}
