% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{coef.gamtree}
\alias{coef.gamtree}
\title{Extract estimated coefficients of a GAM tree}
\usage{
\method{coef}{gamtree}(object, which = "local", ...)
}
\arguments{
\item{object}{an object of class \code{gamtree}.}

\item{which}{character. Specifies whether local (\code{"local"}; 
node-specific estimates from the tree) or globally (\code{"global"})
estimated coefficients should be returned.}

\item{...}{currently not used.}
}
\value{
Returns a matrix (if \code{which = "local"}) or a vector (if 
\code{which = "global"}) with estimated coefficients.
}
\description{
Returns the estimates global or local coefficients of a GAM tree.
}
