% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eco}
\alias{eco}
\title{Example dataset from ecology to illustrate fitting partially additive GAMs}
\format{A data frame with 628 observations and 5 variables}
\usage{
data(eco)
}
\description{
Original dataset was copied from 
https://stackoverflow.com/questions/37037445/using-mob-trees-partykit-package-with-nls-model 
For illustrating the funtionality of package gamtree, two noise variables 
were added: \code{noise} (a continuous variable) and `cluster_id` (a 
categorical indicator for a random intercept).
}
\details{
\itemize{
  \item Species: categorical partitioning variable, indicator for specie (signal).
  \item PAR: continuous predictor variable for the node-specific model (signal).
  \item Pn: continuous response variable.
  \item cluster_id: categorical covariate; indicator for a random intercept
  term (noise).
  \item noise: continuous covariate (noise).
}
}
\examples{
data("eco")
summary(eco)

}
\keyword{datasets}
