% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{gamtree}
\alias{gamtree}
\title{Recursively partition a dataset based on penalized GAMs.}
\usage{
gamtree(
  formula,
  data,
  weights = NULL,
  cluster = NULL,
  offset = NULL,
  verbose = FALSE,
  gam_ctrl = list(),
  mob_ctrl = mob_control(parm = c(1, 2, 4)),
  alt_formula = NULL,
  ...
)
}
\arguments{
\item{formula}{specifies the model formula, consisting of three 
parts: the response variable followed by a tilde ('~'); the terms for the 
node-specific GAMs, followed by a vertical bar ('|') and the potential 
partitioning variables (separated by a '+'). 
The 'by' argument of function
\code{\link[mgcv]{s}} may NOT be used in the node-specific GAM formulation. 
Refrain from
using the dot ('.') to specify all remaining variables in \code{data}, this may yield
unexpected results; make sure to specify each variable in the corresponding part
of the model formula. See Examples.}

\item{data}{\code{data.frame} containing the variables specified in \code{formula}.}

\item{weights}{numeric vector of length \code{nrow(data)}; optional case weights.
A weight of 2, for example, is equivalent to having made exactly the same 
observation twice.}

\item{cluster}{optional numeric or factor vector with a cluster ID to be 
employed for clustered covariances in the parameter stability tests.
This argument should be used when the partitioning variables are not measured
on the individual observation level, but on a higher level. E.g., when 
the response variables consists of repeated measurements on multiple}

\item{offset}{numeric vector of length \code{nrow(data)}. Supplies model 
offset for use in fitting. Note that this offset will always be completely 
ignored when predicting.}

\item{verbose}{logical. Should progress be printed to the commande line in 
every iteration? If true, the iteration number, information on the 
splitting procedure, and the log-likelihood (with df) value of the fitted 
full mixed-effects gam model is printed.}

\item{gam_ctrl}{a list of fit control parameters to replace defaults returned by 
\code{\link[mgcv]{gam.control}}. Values not set assume default values.}

\item{mob_ctrl}{a list with control parameters as returned by 
\code{\link[partykit]{mob_control}} to be passed to function 
\code{\link[partykit]{mob}}. Note: argument `xtype` is set to 
`data.frame`, by default, and cannot be adjusted.}

\item{alt_formula}{list with two elements, for specifying non-standard model formulae
for GAM. E.g., the formula list required for use of the \code{\link[mgcv]{multinom}}
family.}

\item{...}{additional arguments to be passed to function \code{\link[mgcv]{gam}}. 
The same arguments will be employed for the local
(node-specific) GAMs, and the global GAM (if specified in the model \code{formula}).}
}
\value{
Returns an object of class \code{"gamtree"}. This is a list, containing
(amongst others) the GAM-based recursive partition (in \code{$tree}), and the 
fitted full GAM model with local and/or global fitted effects (in \code{$gamm}. 
The following methods are available to extract information from the fitted object:
\code{\link{predict.gamtree}}, for obtaining predicted values for training and new
observations; \code{\link{plot.gamtree}} for plotting the tree and variables' effects; 
\code{\link{coef.gamtree}} for obtaining the estimated coefficients; 
\code{\link{summary.gamtree}} for a summary of the fitted model.
}
\description{
\code{gamtree} recursively partitions a dataset into subgroups with 
penalized GAMs, characterized by differences in the parameter estimates.
}
\examples{
gt <- gamtree(Pn ~ s(PAR, k = 5L) | Species, data = eco, cluster = eco$Specimen)
summary(gt)

}
\seealso{
\code{\link{predict.gamtree}} \code{\link{plot.gamtree}} 
\code{\link{coef.gamtree}} \code{\link{summary.gamtree}}
}
