% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{gamtree}
\alias{gamtree}
\title{Recursively partition a dataset based on a (local) GAM, while accounting for 
globally specified GAM terms.}
\usage{
gamtree(
  formula,
  data,
  weights = NULL,
  cluster = NULL,
  offset = NULL,
  abstol = 0.001,
  verbose = TRUE,
  method = "REML",
  mob_ctrl = NULL,
  ...
)
}
\arguments{
\item{formula}{specifies the model formula, consisting of three or four 
parts: the response variable followed by a tilde, the terms for the 
node-specific GAM followed by a vertical bar, optionally the terms for 
the global GAM followed by a vertical bar, and then the partitioning 
variables. See examples below.}

\item{data}{specifies the dataset (must be a data.frame).}

\item{weights}{currently ignored! optional numeric vector of case weights.}

\item{cluster}{optional numeric or factor vector with a cluster ID to be 
employed for clustered covariances in the parameter stability tests.
This argument should be used when the partitioning variables are not measured
on the individual observation level, but on a higher level. E.g., when 
the response variables consists of repeated measurements on multiple 
observations and the partitioning variables are time-invariant covariates.}

\item{offset}{currently ignored! numeric vector with an a priori known 
component to be included
in the model \code{y ~ x1 + x2 + ....}. Will be applied to both the local
and global model and will not be updated in consecutive iterations.}

\item{abstol}{specifies the convergence criterion. If the log-likelihood 
values of the full mixed-effects gam, from two consecutive iterations differ 
less than abstol, estimation is halted.}

\item{verbose}{logical. Should progress be printed to the commande line in 
every iteration? If true, the iteration number, information on the 
splitting procedure, and the log-likelihood (with df) value of the fitted 
full mixed-effects gam model is printed.}

\item{method}{The smoothing parameter estimation method used by \code{gam()}.
See documentation of function \code{gam()} for details.}

\item{mob_ctrl}{a list with control parameters as returned by \code{mob_control}
to be passed to function \code{mob()}. (Note that argument `xtype` is set to 
`data.frame`, by default, and cannot be adjusted.)}

\item{...}{additional arguments to be passed to the locally fitted 
\code{gam}s, through \code{mob()}'s fitting function (currently, the only 
option is the default \code{gam_fit()}).}
}
\value{
Returns an object of class \code{"gamtree"}. This is a list, containing
(amongst others) the GAM-based recursive partition (in \code{$tree}), and the 
fitted full GAM with both local and/or global fitted effects (in \code{$gamm}).
}
\description{
\code{gamtree} recursively partition a dataset based on a (local) GAM, in
which a smooth function (thin plate regression spline) of a single predictor
variable is estimated, while accounting for globally specified smooth 
functions and random effects.
}
\examples{
## GAM tree without global terms:
gt1 <- gamtree(Pn ~ s(PAR, k = 5L) | Species, data = eco, cluster = eco$specimen, 
              verbose = FALSE)
summary(gt1)
## GAM tree with global terms:
gt2 <- gamtree(Pn ~ s(PAR, k = 5L) | s(cluster_id, bs = "re") + noise | Species, 
              data = eco, cluster = eco$specimen, verbose = FALSE)
summary(gt2)

}
