% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{plot.gamtree}
\alias{plot.gamtree}
\title{Plotting of GAM trees}
\usage{
\method{plot}{gamtree}(x, which = "both", treeplot_ctrl = list(),
  gamplot_ctrl = list(), ...)
}
\arguments{
\item{x}{object of class \code{gamtree}.}

\item{which}{character. The default (\code{"both"}) plots the tree structure, 
followed by the model fitted in the terminal nodes. Alternatively, \code{"tree"}
will plot the tree structure only, and \code{"nodes"} will plot the terminal-node
models only.}

\item{treeplot_ctrl}{list of arguments to be passed to \code{plot.party()}.}

\item{gamplot_ctrl}{list of arguments to be passed to \code{plot,gam()}.}

\item{...}{further arguments, currently not used.}
}
\description{
Takes a fitted GAM tree and plots the smooth functions fitted in each of the
terminal nodes of the tree.
}
\examples{
gt <- gamtree(Pn ~ PAR | Species, data = eco, verbose = FALSE, 
              s_ctrl = list(k = 5L), mob_ctrl = mob_control(maxdepth = 3L))
plot(gt, which = "tree") # fdefault is which = 'both'
plot(gt, which = "nodes", residuals = TRUE)

}
