% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{plot.gamtree}
\alias{plot.gamtree}
\title{Plotting method for GAM trees}
\usage{
\method{plot}{gamtree}(
  x,
  which = "both",
  residuals = TRUE,
  dim = NULL,
  ylim = "firstnode",
  treeplot_ctrl = list(),
  gamplot_ctrl = list(),
  ...
)
}
\arguments{
\item{x}{object of class \code{gamtree}.}

\item{which}{character. The default (\code{"both"}) plots the tree structure, 
followed by the model fitted in the terminal nodes. Alternatively, \code{"tree"}
will plot the tree structure, and \code{"terms"} will plot the smooth (and 
parametric) terms from
the terminal-node-specific and global model. Note that the fitted curves in 
the tree do not convey a conditional function of the predictor on the $x$-axis
(as plotted when "terms" is specified). They are a function of the predictor 
on the $x$-axis, as well as all other predictors in the model and could thus be 
referred to as 'marginal' fitted curves.}

\item{residuals}{Only used when \code{which_terms = local}. Should residuals
be plotted in the node-specific plots?}

\item{dim}{numeric vector of length two. Specifies how many rows and columns 
of plots should be fit on a single page.
\code{NULL} (the default) has the routine leave all settings as they are.
Using \code{par(mfrow = c( , ))} before plotting then provides control over 
the plot dimensions (and number of pages).}

\item{ylim}{\code{"firstplot"} (default), \code{NULL}, or a numeric vector of 
length 2. Only used for plotting the terminal-node models (not the tree). 
Specifies how the limits of the y-axes of the terminal node plots 
should be chosen. The default (\code{"firstnode"}) uses the observations 
in the first node to determine the limits of the y-axes for all plots. 
Alternatively, \code{NULL} will determine the limits of the y-axes 
separately for each plot. Alternatively, a numeric vector of length 
two may be specified, specifying the lower and upper limits of the 
y-axes.}

\item{treeplot_ctrl}{list of (named) arguments to be passed to 
\code{\link[partykit]{plot.party}}.}

\item{gamplot_ctrl}{list of (named) arguments to be passed to 
\code{\link[mgcv]{plot.gam}}. Note that not all arguments 
of \code{plot.gam} are supported. .}

\item{...}{further arguments, currently not used.}
}
\description{
Takes a fitted GAM tree and plots the smooth functions fitted in each of the
terminal nodes of the tree.
}
\section{Warning}{

The plotted terms by default also represent confidence bands. These should
be taken with a big grain of salt, because they do NOT account for the 
searching of the tree structure; they assume the tree structure was
known in advance. They should be interpreted as overly optimistic and with
caution.
}

\examples{
gt <- gamtree(Pn ~ s(PAR, k = 5L) | Species, data = eco, 
               cluster = eco$Specimen) 
plot(gt, which = "tree") # default is which = 'both'
plot(gt, which = "terms")
}
