% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{plot.gamtree}
\alias{plot.gamtree}
\title{Plotting of GAM trees}
\usage{
\method{plot}{gamtree}(x, which = "both", tp_args = list(fitmean =
  FALSE), ...)
}
\arguments{
\item{x}{object of class \code{gamtree}.}

\item{which}{character. The default (\code{"both"}) plots the tree structure, 
followed by the model fitted in the terminal nodes. Alternatively, \code{"tree"}
will plot the tree structure only, and \code{"nodes"} will plot the terminal-node
models only.}

\item{tp_args}{a list with additional control arguments to be passed to
the \code{tp_args} argument of \code{plot.party}.}

\item{...}{arguments to be passed to \code{plot.gam}.}
}
\description{
Takes a fitted GAM tree and plots the smooth functions fitted in each of the
terminal nodes of the tree.
}
