% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{plot.gamtree}
\alias{plot.gamtree}
\title{Plotting method for GAM trees}
\usage{
\method{plot}{gamtree}(
  x,
  which = "both",
  which_terms = "both",
  ylim = "firstnode",
  treeplot_ctrl = list(),
  gamplot_ctrl = list(),
  ...
)
}
\arguments{
\item{x}{object of class \code{gamtree}.}

\item{which}{character. The default (\code{"both"}) plots the tree structure, 
followed by the model fitted in the terminal nodes. Alternatively, \code{"tree"}
will plot the tree structure only, and \code{"nodes"} will plot the smooths from
the terminal-node-specific and global model.}

\item{which_terms}{character; \code{"local"}, \code{"global"} or \code{"both"}.
Specifies whether the local and/or global smooth terms should be plotted.}

\item{ylim}{\code{"firstplot"} (default), \code{NULL}, or a numeric vector of 
length 2. Only used for plotting the terminal-node models (not the tree). 
Specifies how the limits of the y-axes of the terminal node plots 
should be chosen. The default (\code{"firstnode"}) uses the observations 
in the first node to determine the limits of the y-axes for all plots. 
Alternatively, \\code{NULL} will determine the limits of the y-axes 
separately for each plot. Alternatively, a numeric vector of length 
two may be specified, specifying the lower and upper limits of the 
y-axes.}

\item{treeplot_ctrl}{list of (named) arguments to be passed to 
\code{plot.party()}.}

\item{gamplot_ctrl}{list of (named) arguments to be passed to 
\code{plot.gam()}.}

\item{...}{further arguments, currently not used.}
}
\description{
Takes a fitted GAM tree and plots the smooth functions fitted in each of the
terminal nodes of the tree.
}
\examples{
gt <- gamtree(Pn ~ s(PAR, k = 5L) | s(cluster_id, bs = "re") + noise | Species, 
              data = eco, verbose = FALSE, cluster = eco$specimen) 
plot(gt, which = "tree") # fdefault is which = 'both'
plot(gt, which = "nodes", gamplot_ctrl = list(residuals = TRUE))


}
