% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{summary.gamtree}
\alias{summary.gamtree}
\title{Summary method for a fitted GAM tree}
\usage{
\method{summary}{gamtree}(object, ...)
}
\arguments{
\item{object}{object of class \code{gamtree}.}

\item{...}{further arguments to be passed to \code{summary.gam}.}
}
\description{
Prints a summary of the local and/or global terms in a fitted GAM tree.
}
