% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamtree.R
\name{summary.gamtree}
\alias{summary.gamtree}
\title{Summary method for a fitted GAM tree}
\usage{
\method{summary}{gamtree}(object, ...)
}
\arguments{
\item{object}{object of class \code{gamtree}.}

\item{...}{further arguments to be passed to \code{summary.gam}.}
}
\description{
Prints a summary of the local and/or global terms in a fitted GAM tree.
}
\section{Warning}{

The printed results by default also provide standard error and significance
tests. These should be taken with a big grain of salt, because they do NOT 
account for the searching of the tree structure; they assume the tree structure 
was known in advance. They thus should be interpreted as overly optimistic and
with caution.
}

\examples{
## GAM tree without global terms:
gt <- gamtree(Pn ~ s(PAR, k = 5L) | Species, data = eco, cluster = eco$Specimen)
summary(gt)
}
